requirements.txt
# requirements.txt
torch
torchvision
torchaudio
numpy
matplotlib
tqdm
scikit-learn
Pillow
